requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
attrs
certifi
charset-normalizer
colorama
datasets
dill
faiss-gpu
filelock
frozenlist
fsspec
huggingface-hub
idna==3.10
Jinja2
MarkupSafe
mpmath
multidict
multiprocess
networkx
numpy<2
packaging
pandas
peft
propcache
psutil
pyarrow
python-dateutil
pytz
PyYAML
regex
requests
safetensors
sentence_transformers
setuptools
six
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
tzdata
urllib3
xxhash
yarl
